+++ src/hpm_platform/data_import/__init__.py
@@ -5,6 +5,11 @@ from hpm_platform.data_import.calibration_bridge import (
     build_imported_calibration_samples,
     generate_calibration_bridge_report,
 )
+from hpm_platform.data_import.evidence_chain import (
+    evidence_source_chain_ready,
+    generate_evidence_chain_report,
+    load_evidence_chain_config,
+)
 from hpm_platform.data_import.importers import DataImportService, ImportedDataset, inspect_dataset
 from hpm_platform.data_import.model_comparison import generate_model_comparison_report
 from hpm_platform.data_import.vv_audit import generate_external_data_vv_audit
@@ -13,8 +18,11 @@ __all__ = [
     "DataImportService",
     "ImportedDataset",
     "build_imported_calibration_samples",
+    "evidence_source_chain_ready",
     "generate_calibration_bridge_report",
+    "generate_evidence_chain_report",
     "generate_external_data_vv_audit",
     "generate_model_comparison_report",
     "inspect_dataset",
+    "load_evidence_chain_config",
 ]